Separating AIE kernel related code
