+++ src/ggml-hsa/ggml-hsa.cpp
@@ -336,10 +336,7 @@ static ggml_hsa_device_info & ggml_hsa_info_mut() {
 
 const ggml_hsa_device_info & ggml_hsa_info() { return ggml_hsa_info_mut(); }
 
-/**
- * @brief Returns the device info associated with @p device_id.
- */
-static const ggml_hsa_device_info::device_info & ggml_hsa_get_device_info(std::int32_t device_id) {
+const ggml_hsa_device_info::device_info & ggml_hsa_get_device_info(std::int32_t device_id) {
     const auto & info = ggml_hsa_info();
     const auto & dev_info = info.devices[device_id];
     return dev_info;
@@ -348,14 +345,12 @@ static const ggml_hsa_device_info::device_info & ggml_hsa_get_device_info(std::i
 /**
  * @brief Caches the @p new_kernel for the tensor @p tensor.name on the device @p device_id.
  */
-static std::shared_ptr<ggml_hsa_kernel>
-ggml_hsa_cache_kernel(std::string kernel_name, std::int32_t device_id, ggml_hsa_aie_kernel kernel) {
+static std::shared_ptr<ggml_hsa_kernel> ggml_hsa_cache_kernel(
+    std::string kernel_name, std::int32_t device_id, std::shared_ptr<ggml_hsa_kernel> kernel) {
     auto & info = ggml_hsa_info_mut();
     auto & dev_info = info.devices[device_id];
     auto & kernels = dev_info.kernels;
-    auto result = kernels.emplace(std::move(kernel_name),
-                                  std::static_pointer_cast<ggml_hsa_kernel>(
-                                      std::make_shared<ggml_hsa_aie_kernel>(std::move(kernel))));
+    auto result = kernels.emplace(std::move(kernel_name), std::move(kernel));
     if (!result.second) {
         GGML_ABORT("%s: kernel %s already exists on device %d\n", __func__, kernel_name.c_str(),
                    device_id);
@@ -531,8 +526,7 @@ ggml_backend_hsa_tensor_extra::ggml_backend_hsa_tensor_extra(
         kernel = ggml_hsa_get_cached_kernel(kernel_name, dev_info.device);
         // retrieve a cached kernel or create a new one
         if (kernel == nullptr) {
-            ggml_hsa_aie_kernel aie_kernel;
-            if (ggml_hsa_create_aie_kernel(dev_info, kernel_name, tensor, aie_kernel) !=
+            if (ggml_hsa_create_kernel(dev_info, kernel_name, tensor, kernel) !=
                 GGML_STATUS_SUCCESS) {
                 throw std::runtime_error{std::string{"Failed to create kernel for tensor \""}
                                              .append(tensor.name)
@@ -540,8 +534,8 @@ ggml_backend_hsa_tensor_extra::ggml_backend_hsa_tensor_extra(
                                              .append(ggml_op_desc(&tensor))
                                              .append(")")};
             }
-            kernel = ggml_hsa_cache_kernel(std::move(kernel_name), dev_info.device,
-                                           std::move(aie_kernel));
+            kernel =
+                ggml_hsa_cache_kernel(std::move(kernel_name), dev_info.device, std::move(kernel));
         }
     }
 }
@@ -631,104 +625,6 @@ ggml_backend_hsa_context::~ggml_backend_hsa_context() {
 }
 
 void ggml_backend_hsa_context::free_pending_payloads() { pending_payloads.clear(); }
-
-/**
- * @brief Dispatches an HSA packet.
- *
- * @todo @p dispatch_signal is not used yet.
- *
- * @param[in] queue queue to enqueue the packet
- * @param[in] signal signal to notify for packet completion
- * @param[in] payload packet payload
- * @param[in] payload_size payload size in dwords
- */
-static void ggml_hsa_dispatch_packet(hsa_queue_t * queue,
-                                     hsa_signal_t /* dispatch_signal */,
-                                     hsa_amd_aie_ert_start_kernel_data_t * payload,
-                                     std::size_t payload_size) {
-    hsa_amd_aie_ert_packet_t pkt{};
-    pkt.header.header = HSA_PACKET_TYPE_VENDOR_SPECIFIC << HSA_PACKET_HEADER_TYPE;
-    pkt.header.AmdFormat = HSA_AMD_PACKET_TYPE_AIE_ERT;
-    pkt.state = HSA_AMD_AIE_ERT_STATE_NEW;
-    pkt.count = payload_size;
-    pkt.opcode = HSA_AMD_AIE_ERT_START_CU;
-    pkt.payload_data = reinterpret_cast<std::uint64_t>(payload);
-    // TODO add pkt->completion_signal = dispatch_signal
-
-    const std::uint64_t wr_idx = hsa_queue_add_write_index_relaxed(queue, 1);
-    const std::uint64_t packet_id = wr_idx % queue->size;
-    *(static_cast<hsa_amd_aie_ert_packet_t *>(queue->base_address) + packet_id) = pkt;
-
-    hsa_signal_store_screlease(queue->doorbell_signal, wr_idx);
-}
-
-/**
- * @brief Decomposes a 64-bit address to two @c std::uint32_t.
- */
-inline std::tuple<std::uint32_t, std::uint32_t> ggml_hsa_addr_to_hilo(const void * address) {
-    static_assert(sizeof(void *) == 2 * sizeof(std::uint32_t));
-    return {reinterpret_cast<uint64_t>(address) >> 32,
-            reinterpret_cast<uint64_t>(address) & 0xFFFFFFFF};
-}
-
-ggml_status ggml_hsa_aie_kernel::dispatch(ggml_backend_hsa_context & ctx,
-                                          ggml_tensor * src_tensors[],
-                                          std::size_t num_src_tensors,
-                                          ggml_tensor * dst_tensor) {
-    const auto & dev_info = ggml_hsa_get_device_info(ctx.device);
-    const std::size_t packet_dwords =
-        3 /* instructions */ + (num_src_tensors + 1) * 3 /* source and destination tensors */;
-    void * ptr = nullptr;
-    if (auto status =
-            hsa_amd_memory_pool_allocate(dev_info.kernarg_memory.memory_pool, 64, 0, &ptr);
-        status != HSA_STATUS_SUCCESS) {
-        GGML_LOG_ERROR("%s: failed to allocate hsa_queue packet storage (%s)\n", __func__,
-                       ggml_hsa_get_status_string(status));
-        return GGML_STATUS_ALLOC_FAILED;
-    }
-    ctx.pending_payloads.emplace_back(ptr);
-
-    auto cmd_payload = static_cast<hsa_amd_aie_ert_start_kernel_data_t *>(ptr);
-    cmd_payload->pdi_addr =
-        const_cast<void *>(static_cast<const void *>(pdi.data())); // PDI to use with this command
-
-    // transaction opcode; not counted in packet_dwords
-    cmd_payload->data[0] = 0x3;
-    cmd_payload->data[1] = 0x0;
-
-    std::size_t dword_idx = 2;
-
-    // instructions; 3 dwords
-    std::tie(cmd_payload->data[dword_idx + 1], cmd_payload->data[dword_idx]) =
-        ggml_hsa_addr_to_hilo(insts.data());
-    cmd_payload->data[dword_idx + 2] = static_cast<std::uint32_t>(insts.size());
-
-    dword_idx += 3;
-
-    // sources; 2 dwords each
-    for (std::size_t src_idx = 0; src_idx < num_src_tensors; ++src_idx, dword_idx += 2) {
-        std::tie(cmd_payload->data[dword_idx + 1], cmd_payload->data[dword_idx]) =
-            ggml_hsa_addr_to_hilo(src_tensors[src_idx]->data);
-    }
-
-    // destination; 2 dwords
-    std::tie(cmd_payload->data[dword_idx + 1], cmd_payload->data[dword_idx]) =
-        ggml_hsa_addr_to_hilo(dst_tensor->data);
-
-    dword_idx += 2;
-
-    // sizes; 1 dword per tensor
-    for (std::size_t src_idx = 0; src_idx < num_src_tensors; ++src_idx, ++dword_idx) {
-        cmd_payload->data[dword_idx] = ggml_nbytes(src_tensors[src_idx]);
-    }
-    cmd_payload->data[dword_idx] = ggml_nbytes(dst_tensor);
-
-    assert(dword_idx == packet_dwords + 1); // 2 extra uncounted dwords
-
-    ggml_hsa_dispatch_packet(ctx.queue, ctx.dispatch_signal, cmd_payload, packet_dwords);
-
-    return GGML_STATUS_SUCCESS;
-}
 
 void ggml_hsa_wait_dispatches(ggml_backend_hsa_context & ctx) {
     if (auto val = hsa_signal_wait_scacquire(ctx.dispatch_signal, HSA_SIGNAL_CONDITION_EQ, 0,
